edge_glass/requirements.txt
# ============================================================
# SDPRS Edge Glass Node - Python Dependencies (M1 MVP)
# Smart Disaster Prevention Response System
# ============================================================
# 安裝: pip install -r requirements.txt
# ============================================================

# ------------------------------------------------------------
# 核心依賴
# ------------------------------------------------------------

# 影像處理 — headless 版本（無 GUI，適合 Pi）
opencv-python-headless>=4.8.0,<5.0.0

# 數值計算（影像陣列、音訊 FFT）
numpy>=1.24.0,<2.0.0

# YAML 配置檔載入
PyYAML>=6.0,<7.0

# 音訊擷取（PyAudio — 需系統先安裝 portaudio 和 python3-dev）
# sudo apt install portaudio19-dev python3-dev
# deploy_sync.sh init-glass 會自動安裝這些系統依賴
PyAudio>=0.2.13

# ------------------------------------------------------------
# M2 通訊模組依賴
# ------------------------------------------------------------

# MQTT 客戶端 — 心跳發布 + 指令訂閱（Thread 3）
paho-mqtt>=2.0.0,<3.0.0

# HTTP 客戶端 — JSON/MP4 上傳 + 快照推送（Thread 4, 5）
httpx>=0.25.0,<1.0.0

# 系統監控 — CPU 溫度、記憶體使用率（心跳 payload + 熱管理 Thread 6）
psutil>=5.9.0,<6.0.0

# ------------------------------------------------------------
# 開發 / 測試
# ------------------------------------------------------------
pytest>=7.0.0
pytest-cov>=4.0.0